# src/swagger_documentation.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "DigitalStore API",
    "description": "Documentação da API da digital store"
  },
  "host": "https://digital-store-backend-c4an.onrender.com",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "JWT": {
      "description": "JWT token",
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/carrinhos/": {
      "get": {
        "tags": [
          "Carrinhos"
        ],
        "description": "Retorna lista de carrinhos.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna lista de carrinhos",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "carrinho_id": {
                    "type": "number",
                    "example": 1
                  },
                  "usuario_id": {
                    "type": "number",
                    "example": 1
                  },
                  "itens": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "carrinho_id": {
                          "type": "number",
                          "example": 1
                        },
                        "produto_id": {
                          "type": "number",
                          "example": 1
                        },
                        "quantidade": {
                          "type": "number",
                          "example": 1
                        }
                      }
                    }
                  },
                  "criado_em": {
                    "type": "string",
                    "example": "2024-08-26 23:31:00"
                  },
                  "atualizado_em": {
                    "type": "string",
                    "example": "2024-08-26 23:31:00"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "detail": {
                  "type": "string",
                  "example": "Usuário não encontrado."
                },
                "severity": {
                  "type": "string",
                  "example": "warm"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 422
                },
                "detail": {
                  "type": "string",
                  "example": "mensagem do sistema"
                },
                "severity": {
                  "type": "string",
                  "example": "danger"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Carrinhos"
        ],
        "description": "Cria um carrinho",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "usuario_id": {
                  "type": "number",
                  "example": 1
                },
                "itens": {
                  "type": "array",
                  "example": [],
                  "items": {}
                }
              },
              "required": [
                "usuario_id",
                "itens"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Carrinho criado",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "detail": {
                  "type": "string",
                  "example": "Carrinho criado"
                },
                "severity": {
                  "type": "string",
                  "example": "success"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "detail": {
                  "type": "string",
                  "example": "Usuário não encontrado."
                },
                "severity": {
                  "type": "string",
                  "example": "warm"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 422
                },
                "detail": {
                  "type": "string",
                  "example": "mensagem do sistema"
                },
                "severity": {
                  "type": "string",
                  "example": "danger"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Carrinhos"
        ],
        "description": "Edita um carrinho.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "carrinho_id": {
                  "type": "number",
                  "example": 1
                },
                "usuario_id": {
                  "type": "number",
                  "example": 1
                },
                "itens": {
                  "type": "array",
                  "example": [],
                  "items": {}
                }
              },
              "required": [
                "carrinho_id",
                "usuario_id",
                "itens"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Carrinho atualizado",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "detail": {
                  "type": "string",
                  "example": "Carrinho atualizado"
                },
                "severity": {
                  "type": "string",
                  "example": "success"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "detail": {
                  "type": "string",
                  "example": "Usuário não encontrado."
                },
                "severity": {
                  "type": "string",
                  "example": "warm"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 422
                },
                "detail": {
                  "type": "string",
                  "example": "mensagem do sistema"
                },
                "severity": {
                  "type": "string",
                  "example": "danger"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/carrinhos/{id}": {
      "get": {
        "tags": [
          "Carrinhos"
        ],
        "description": "Retorna um carrinho.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna um carrinho.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "carrinho_id": {
                    "type": "number",
                    "example": 1
                  },
                  "usuario_id": {
                    "type": "number",
                    "example": 1
                  },
                  "itens": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "carrinho_id": {
                          "type": "number",
                          "example": 1
                        },
                        "produto_id": {
                          "type": "number",
                          "example": 1
                        },
                        "quantidade": {
                          "type": "number",
                          "example": 1
                        }
                      }
                    }
                  },
                  "criado_em": {
                    "type": "string",
                    "example": "2024-08-26 23:31:00"
                  },
                  "atualizado_em": {
                    "type": "string",
                    "example": "2024-08-26 23:31:00"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "detail": {
                  "type": "string",
                  "example": "Usuário não encontrado."
                },
                "severity": {
                  "type": "string",
                  "example": "warm"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 422
                },
                "detail": {
                  "type": "string",
                  "example": "mensagem do sistema"
                },
                "severity": {
                  "type": "string",
                  "example": "danger"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Carrinhos"
        ],
        "description": "Deleta um carrinho.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Carrinho deletado.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "detail": {
                  "type": "string",
                  "example": "Carrinho deletado."
                },
                "severity": {
                  "type": "string",
                  "example": "success"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "detail": {
                  "type": "string",
                  "example": "Usuário não encontrado."
                },
                "severity": {
                  "type": "string",
                  "example": "warm"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 422
                },
                "detail": {
                  "type": "string",
                  "example": "mensagem do sistema"
                },
                "severity": {
                  "type": "string",
                  "example": "danger"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  },
...
}